GUI Update!
Added a barebones GUI accessed using the --gui flag on startup. You can select color, and it automatically assumes that you want to use the working detection algorithm and connect over network tables.

This makes it much easier to start the vision code. The next step is to create a script that starts the vision code and the driver station. I could also potentially embed the camera feed into the Tkinter GUI to make it even simpler.