fairly major code changes
made it so the vision code actually gets an input (finally) and made it so the top of the vision code's camera preview displays red, yellow, and green for the different stages of auto orient (red means not oriented, yellow means currently orienting, and green means fully oriented/ready to shoot).

We should be fully ready for tomorrow's competition!